chore(build): Update Gradle file hashes
This commit updates the Gradle `fileHashes.lock` file. This file is automatically generated and managed by the Gradle build system to support incremental builds by tracking the hashes of project files.

This change is a side-effect of other modifications to source code or build configurations and ensures the build cache is up-to-date.